Created a layer between Bootstrapper<T> and BootstrapperBase to allow the situation that no root view is needed.
